New Features ------------
1) Compound time signatures with internal subdivision structure. It is
possible to define time signatures as a sum of parts (3/8+2/8), with
heterogeneous denominators (4/4+1/16) and also to define the internal
subdivisions: for example, to define a 7/8 subdivided as 2-3-2, use
7/8(2-3-2) in the scorestruct definition or a TimeSignature object
of the form TimeSignature((7, 8), subdivisionStruct=(2, 3, 2))

Internal
--------

1) new IndentedWriter class, handles all text output which needs
indentation. This is the case of lilypond output at the time, since
musicxml is output not directly as text but via the standard lib.

2) Removed some code not being used. (centsshown, which was part
of maelzel.core and is now delayed to the scoring step

Test
----

1) Added a set of musicxml files with rendered pdfs for testing

Documentation
-------------

* new tutorial: clicktrack
* updated many demos / tutorials, to bring them up to date